requirements.txt
ann-visualizer==2.5
Keras==2.4.3
Keras-Preprocessing==1.1.0
numpy==1.18.1
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
openpyxl==3.0.3
pickleshare==0.7.5
Pillow==7.0.0
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex==2020.9.27
scikit-learn==0.23.2
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///tmp/build/80754af9/tensorflow-base_1597847441277/work/tensorflow_pkg/tensorflow-2.3.0-cp37-cp37m-linux_x86_64.whl
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
